backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==41.0.7

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database connectors (for schema analysis)
pymysql==1.1.0
psycopg2-binary==2.9.9
pymssql==2.2.11

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
